[ORO-0] Make Orochi premake independent
