+++ Test/DeviceEnum/premake5.lua
@@ -4,6 +4,5 @@ project "DeviceEnum"
     location "%{wks.location}/Test/%{prj.name}"
 
     useOrochi()
-    linkVersionLib()
 
     files { "*.cpp" }
