+++ Test/Texture/premake5.lua
@@ -4,7 +4,6 @@ project "Texture"
     location "%{wks.location}/Test/%{prj.name}"
 
     useOrochi()
-    linkVersionLib()
 
     files { "../../UnitTest/contrib/stb/**.h", "../../UnitTest/contrib/stb/**.cpp" }
     files { "texture_test_kernel.hpp", "*.cpp" }
