+++ Test/VulkanComputeSimple/premake5.lua
@@ -4,7 +4,6 @@ project "VulkanComputeSimple"
     location "%{wks.location}/Test/%{prj.name}"
 
     useOrochi()
-    linkVersionLib()
     linkWin32SystemLibs()
     includedirs { "./" }
     files { "*.cpp" }
