+++ Test/WMMA/premake5.lua
@@ -4,7 +4,6 @@ project "WMMA"
     location "%{wks.location}/Test/%{prj.name}"
 
     useOrochi()
-    linkVersionLib()
 
     files { "*.h", "*.cpp" }
     files { "half.hpp" }
